Extract <DiscussionPostForm />  component
